requirements.txt
arcade==2.6.17
pillow==9.3.0